fix: move out from sub dirs
