+++ packages/opentelemetry-propagator-jaeger/src/JaegerHttpTraceFormat.ts
